requirements.txt
Django==3.1.3
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0
drf-yasg==1.20.0